#27 - moves settings files to its module
