+++ app/src/main/res/values/strings.xml
@@ -132,8 +132,6 @@
     <string name="recurring_expense_billing_day">Dia</string>
     <string name="recurring_total_title">Total:</string>
 
-    <string name="settings_dark_mode">Modo escuro</string>
-
     <string name="date_picker_title">Selecione uma data</string>
     <string name="date_picker_label">Data da operação</string>
     <string name="date_picker_cancel">Cancelar</string>
